begin pen and touch

+++ sources/Input/Input/Implementations/SDL3/Devices/ISdlDevice.cs
@@ -9,6 +9,6 @@ namespace Silk.NET.Input.SDL3;
 /// <typeparam name="T"></typeparam>
 internal interface ISdlDevice<out T> : IInputDevice where T : SdlDevice
 {
-    public static abstract T? CreateDevice(uint sdlDeviceId, SdlInputBackend backend);
+    public static abstract T? CreateDevice(ulong sdlDeviceId, SdlInputBackend backend);
 
 }
